Ajustement BarreNavigation + refactorisation quelques composants.
