+++ liste-repertoire-frontend/src/pages/PageModifierDemandeSpeciale.js
@@ -3,9 +3,7 @@ import FormulaireModifierDemandeSpeciale from "../composants/FormulaireModifierD
 import Button from 'react-bootstrap/Button'
 import { Link } from 'react-router-dom';
 
-// A FINIR
-function PageModifierDemandeSpeciale({ match })
-{
+function PageModifierDemandeSpeciale({ match }) {
     const id = match.params.id;
     return (
         <>
